backend/requirements.txt
Django~=5.0.4
djangorestframework~=3.15.1
django-filter~=24.2
psycopg2~=2.9.9
drf-spectacular~=0.27.2
PyJWT~=2.8.0
Pillow~=10.3.0
django-cors-headers~=4.3.1
django-extensions~=3.2.3
Werkzeug~=3.0.3
pyOpenSSL~=24.1.0
gunicorn~=22.0.0
